Predicting the quality of Tea
In this project, I created a dataset myself to train the model so that it predicts the quality of a tea sample given a set attributes.
Since our project is a classification problem, we need to use one of the classification algorithms out of many. To choose the best algorithm for our project, we test the accuracy of each algorithm in relation to our project. We find that Linear Discriminant Analysis gives us the maximum accuracy of 95%.
Finally, we get three different clusters of types of tea perfectly separated using the eigen vectors for each attribute of the types of tea and plotting it on the graph. So when a new sample of tea is introduced, the machine knows how to classify it as the type of tea it is and places it into one of the three clusters.